requirements.txt
PyYAML
flask
langchain-ollama
langchain-community
chromadb
sympy
TTS>=0.22.0
torch
numpy<2.0.0
pygame
duckduckgo-search
edge-tts
werkzeug
requests
opencv-python
playwright
pyttsx3
google-genai
python-dotenv
